+++ codex-rs/codex-mcp/src/client.rs
@@ -8,13 +8,13 @@ use std::sync::atomic::Ordering;
 use std::time::Duration;
 use std::time::Instant;
 
+use crate::elicitation::ElicitationRequestManager;
 use crate::mcp::CODEX_APPS_MCP_SERVER_NAME;
 use crate::mcp::ToolPluginProvenance;
 use crate::mcp_connection::CachedCodexAppsToolsLoad;
 use crate::mcp_connection::CodexAppsToolsCacheContext;
 use crate::mcp_connection::DEFAULT_STARTUP_TIMEOUT;
 use crate::mcp_connection::DEFAULT_TOOL_TIMEOUT;
-use crate::mcp_connection::ElicitationRequestManager;
 use crate::mcp_connection::MCP_SANDBOX_STATE_META_CAPABILITY;
 use crate::mcp_connection::McpRuntimeEnvironment;
 use crate::mcp_connection::ToolFilter;
