+++ codex-rs/codex-mcp/src/lib.rs
@@ -39,6 +39,7 @@ pub use mcp_connection::declared_openai_file_input_param_names;
 pub use mcp_connection::filter_non_codex_apps_mcp_tools_only;
 
 pub(crate) mod client;
+pub(crate) mod elicitation;
 pub(crate) mod manager;
 pub(crate) mod mcp;
 pub(crate) mod mcp_connection;
