+++ codex-rs/codex-mcp/src/manager.rs
@@ -11,11 +11,11 @@ use crate::client::MCP_TOOLS_LIST_DURATION_METRIC;
 use crate::client::ManagedClient;
 use crate::client::StartupOutcomeError;
 use crate::client::list_tools_for_client_uncached;
+use crate::elicitation::ElicitationRequestManager;
 use crate::mcp::CODEX_APPS_MCP_SERVER_NAME;
 use crate::mcp::ToolPluginProvenance;
 use crate::mcp_connection::CodexAppsToolsCacheContext;
 use crate::mcp_connection::CodexAppsToolsCacheKey;
-use crate::mcp_connection::ElicitationRequestManager;
 use crate::mcp_connection::McpRuntimeEnvironment;
 use crate::mcp_connection::ToolInfo;
 use crate::mcp_connection::emit_duration;
