+++ codex-rs/codex-mcp/src/mcp_connection_manager_tests.rs
@@ -4,12 +4,12 @@ use crate::client::ManagedClient;
 use crate::client::StartupOutcomeError;
 use crate::client::elicitation_capability_for_server;
 use crate::declared_openai_file_input_param_names;
+use crate::elicitation::ElicitationRequestManager;
+use crate::elicitation::elicitation_is_rejected_by_policy;
 use crate::mcp_connection::CODEX_APPS_TOOLS_CACHE_SCHEMA_VERSION;
 use crate::mcp_connection::CodexAppsToolsCacheContext;
-use crate::mcp_connection::ElicitationRequestManager;
 use crate::mcp_connection::ToolFilter;
 use crate::mcp_connection::ToolInfo;
-use crate::mcp_connection::elicitation_is_rejected_by_policy;
 use crate::mcp_connection::filter_tools;
 use crate::mcp_connection::load_startup_cached_codex_apps_tools_snapshot;
 use crate::mcp_connection::read_cached_codex_apps_tools;
